refactor(sessions): extract form constants & types; centralize responsive view util
