+++ src/modules/sessions/const/session-form.const.ts
@@ -1,0 +1,12 @@
+import type { SessionFormState } from "../types/session-form.type";
+
+export const INITIAL_SESSION_FORM: SessionFormState = {
+  title: "",
+  description: "",
+  startDate: "",
+  startTime: "",
+  endTime: "",
+  linkedSkill: "",
+  linkedQuest: "",
+  color: "#3B82F6",
+};
